src/agent_c_tools/pyproject.toml
[project]
name = "agent_c-tools"
version = "0.1.3"
description = "Reference tools / tool interfaces for Agent C"
dependencies = [
    "agent_c-core>=0.1.3", # Includes pydantic as a dependency
    "pandas==2.2.3", # googletrends, dynamics (overlap agent_c_demo)
    "pyarrow==17.0.0", # googletrends, # data_vis/dataframe
    "mermaid-py==0.6.0", # mermaid
    "feedparser==6.0.11", # rss
    "readability-lxml==0.8.1", # web
    "markdownify==0.13.1", # web
    "selenium==4.25.0", # web, dynamics
    "httpx>=0.27", # weather, web
    "duckduckgo_search==7.5.5", # web_search-duckduckgo
    "google-search-results==2.4.2", # web_search-google_serp
    "pytrends==4.9.2", # web_search-google_trends
    "requests==2.32.3", # web_search-hacker_news, Youtube_Comments/Search
    "newsapi-python==0.2.7", # web_search-news_api
    "beautifulsoup4==4.12.3", # web_search-seeking_alpha
    "aiohttp==3.10.10", # web_search-seeking_alpha
    "tavily-python==0.5.0", # web_search-tavily
    "wikipedia==1.4.0", # web_search-wikipedia
    "lxml_html_clean",
    "aiobotocore>=2.21.1",
    "azure-storage-blob>=12.25.1",
    "azure-core>=1.32.0",
    "unidiff>=0.7.5",
    "markdown", # markdown html report tool
    "beautifulsoup4==4.12.3", # dynamics (overlap agent_c_tools)
    "pytz==2024.2", # dynamics
    "lxml==5.3.0", # dynamics (overlaps agent_c_voice)
    "lxml_html_clean==0.2.2", # Dynamics (overlaps agent_c_voice)
    "openpyxl==3.1.5", # dynamics
    "mcp>=0.7.0", # MCP SDK
    "pyyaml>=6.0", # Configuration file support
    "python-docx==1.1.2", # markdown html report tool
    "sqlparse",
    "mysql-connector-python",
    "pytest-playwright",
    "sympy",
    "numpy",
    "scipy",
    "matplotlib",
    "pdfplumber",
    "matplotlib==3.9.2",    # data_vis
    "seaborn==0.13.2",      # data_vis

]

requires-python = ">=3.10"
authors = [
    { name = "Donavan Stanley", email = "donavan.stanley@centricconsulting.com" },
    { name = "Joe Ours", email = "joesph.ours@centricconsulting.com" },
    { name = "Sushant Bhatnagar", email = "sushant.bhatnagar@centricconsulting.com" }
]
maintainers = [
    { name = "Donavan Stanley", email = "donavan.stanley@centricconsulting.com" },
]

readme = "README.MD"
license = { text = "Business Source License 1.1" }
keywords = ["gpt", "claude", "agents", "ai", "centric"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python"
]

[project.urls]
Homepage = "https://github.com/centricconsulting/agent_c_framework"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"


